Revert "Merge pull request #424 from adamtheturtle/separate-update"
This reverts commit cf1e68799a249cbdfbd0769f7758c113f22c4bfe, reversing
changes made to bde77f6a509ea44d1f10a81e519af2b318fbb64e.
